+++ scvelo/core/tests/test_linear_models.py
@@ -13,7 +13,7 @@ from scvelo.core import LinearRegression
 class TestLinearRegression:
     @given(
         x=arrays(
-            np.float,
+            float,
             shape=st.integers(min_value=1, max_value=100),
             elements=st.floats(
                 min_value=-1e3, max_value=1e3, allow_infinity=False, allow_nan=False
@@ -33,7 +33,7 @@ class TestLinearRegression:
 
     @given(
         x=arrays(
-            np.float,
+            float,
             shape=st.tuples(
                 st.integers(min_value=1, max_value=100),
                 st.integers(min_value=1, max_value=100),
@@ -43,7 +43,7 @@ class TestLinearRegression:
             ),
         ),
         coef=arrays(
-            np.float,
+            float,
             shape=100,
             elements=st.floats(
                 min_value=-1000, max_value=1000, allow_infinity=False, allow_nan=False
